Gameplay Update And Experiments
Many experiments, and also added the main core idea for the game into my brain. trying to figure out a way to get the impact point of the collision and also get a mirror vector to impact of bullets. Equation is already there just need to calculate the right normals